Add error checking with result of ft_atof

+++ includes/types.h
@@ -43,7 +43,7 @@ typedef struct s_camera
 {
 	t_point	view_point;
 	t_vec3	orientation;
-	int		fov;
+	double	fov;
 }	t_camera;
 
 typedef struct s_light
